refactor(LineReponse): SectionResponse 삭제
